requirements.txt
django>=4.2.11
djangorestframework>=3.15.1
djangorestframework-jsonapi>=6.1.0
django-filter>=24.2
git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
supervision>=0.19.0
ultralytics>=8.1.47